extract spec_info
