+++ src/spec.h
@@ -461,11 +461,8 @@ int IndexSpec_CheckAllowSlopAndInorder(const IndexSpec *sp, t_fieldMask fm, Quer
  */
 const FieldSpec *IndexSpec_GetFieldBySortingIndex(const IndexSpec *sp, uint16_t idx);
 
-/* Initialize some index stats that might be useful for scoring functions */
-void IndexSpec_GetStats(IndexSpec *sp, RSIndexStats *stats);
-
-/* Get the number of indexing failures */
-size_t IndexSpec_GetIndexErrorCount(const IndexSpec *sp);
+// IndexSpec_GetStats, IndexSpec_GetIndexErrorCount moved to spec_info.h
+#include "spec_info.h"
 
 /*
  * Parse an index spec from redis command arguments.
@@ -530,14 +527,7 @@ void IndexesScanner_Cancel(struct IndexesScanner *scanner);
 void IndexesScanner_ResetProgression(struct IndexesScanner *scanner);
 
 void IndexSpec_ScanAndReindex(RedisModuleCtx *ctx, StrongRef ref);
-/**
- * Exposing all the fields of the index to INFO command.
- * @param ctx - the redis module info context
- * @param sp - the index spec
- * @param obfuscate - if true, obfuscate the index name and field names
- * @param skip_unsafe_ops - if true, skips operations unsafe in signal handler context (allocations, locks)
- */
-void IndexSpec_AddToInfo(RedisModuleInfoCtx *ctx, IndexSpec *sp, bool obfuscate, bool skip_unsafe_ops);
+// IndexSpec_AddToInfo moved to spec_info.h
 
 /**
  * Gets the next text id from the index. This does not currently
@@ -651,36 +641,7 @@ typedef struct DebugIndexesScanner {
 } DebugIndexesScanner;
 
 
-double IndexesScanner_IndexedPercent(RedisModuleCtx *ctx, IndexesScanner *scanner, const IndexSpec *sp);
-
-/**
- * @return the overhead used by the TAG fields in `sp`, i.e., the size of the
- * TrieMaps used for the `values` and `suffix` fields.
- */
-size_t IndexSpec_collect_tags_overhead(const IndexSpec *sp);
-
-/**
- * @return the overhead used by the TEXT fields in `sp`, i.e., the size of the
- * sp->terms and sp->suffix Tries.
- */
-size_t IndexSpec_collect_text_overhead(const IndexSpec *sp);
-
-/**
- * @return the overhead used by the NUMERIC and GEO fields in `sp`, i.e., the accumulated size of all
- * numeric tree structs.
- */
-size_t IndexSpec_collect_numeric_overhead(IndexSpec *sp);
-
-/**
- * @return all memory used by the index `sp`.
- * Uses the sizes of the doc-table, tag and text overhead if they are not `0`
- * (otherwise compute them in-place). Vector overhead is expected to be passed in as an argument
- * and will not be computed in-place
- * TODO: fIx so this will account for the entire index memory, preferably by using an allocator,
- * currently it is a best effort that account only for part of the actual memory.
- */
-size_t IndexSpec_TotalMemUsage(IndexSpec *sp, size_t doctable_tm_size, size_t tags_overhead,
-  size_t text_overhead, size_t vector_overhead);
+// IndexesScanner_IndexedPercent, collect_*_overhead, TotalMemUsage moved to spec_info.h
 
 /**
 * obfuscate argument is used to determine how we will format the index name
